requirements.txt
# Core dependencies
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0

# Document processing
python-docx==0.8.11
PyPDF2==3.0.1
python-multipart==0.0.6

# NLP libraries
transformers==4.33.2
torch==2.0.1
spacy==3.6.1
nltk==3.8.1

# UI
streamlit==1.26.0
gradio==3.40.1

# Development tools
pytest==7.4.1
black==23.7.0
isort==5.12.0
flake8==6.1.0

# Deployment
python-dotenv==1.0.0